remove host SBOM collector from collector component
